+++ packages/article-flag/article-flag.showcase.js
@@ -1,5 +1,6 @@
 import React from "react";
 import ArticleFlag, {
+  ArticleFlags,
   NewArticleFlag,
   UpdatedArticleFlag,
   ExclusiveArticleFlag,
@@ -36,6 +37,11 @@ export default {
     {
       component: () => <NewArticleFlag color="blue" />,
       name: "Article Flag with text colour",
+      type: "story"
+    },
+    {
+      component: () => <ArticleFlags flags={["UPDATED", "EXCLUSIVE"]} />,
+      name: "Article Flags",
       type: "story"
     }
   ],