+++ packages/article-magazine-comment/src/styles/shared.js
@@ -1,4 +1,7 @@
-import styleguide from "@times-components/styleguide";
+import styleguide, {
+  tabletRowPadding,
+  tabletWidth
+} from "@times-components/styleguide";
 
 const { colours, fontFactory, spacing } = styleguide();
 const sharedStyles = {
@@ -21,6 +24,10 @@ const sharedStyles = {
     marginRight: spacing(2),
     paddingTop: spacing(7)
   },
+  containerTablet: {
+    alignSelf: "center",
+    width: tabletWidth
+  },
   datePublication: {
     ...fontFactory({
       font: "supporting",
@@ -29,19 +36,19 @@ const sharedStyles = {
     color: colours.functional.secondary,
     marginTop: spacing(3)
   },
-  flag: {
-    flexDirection: "row",
+  flags: {
     marginBottom: spacing(3),
     marginTop: spacing(1)
-  },
-  flagContainer: {
-    marginRight: spacing(3)
   },
   label: {
     marginBottom: spacing(2)
   },
   leadAssetContainer: {
     marginBottom: spacing(4)
+  },
+  leadAssetContainerTablet: {
+    alignSelf: "center",
+    width: tabletWidth - tabletRowPadding
   },
   meta: {
     flexDirection: "row",