+++ packages/article-main-comment/src/styles/shared.js
@@ -1,4 +1,4 @@
-import styleguide from "@times-components/styleguide";
+import styleguide, { tabletWidth } from "@times-components/styleguide";
 
 const { colours, fontFactory, spacing } = styleguide();
 const sharedStyles = {
@@ -16,13 +16,12 @@ const sharedStyles = {
   },
   container: {
     alignItems: "center",
-    borderBottomColor: colours.functional.keyline,
-    borderBottomWidth: 1,
-    marginBottom: spacing(4),
-    marginLeft: spacing(2),
-    marginRight: spacing(2),
-    paddingBottom: spacing(5),
-    paddingTop: spacing(7)
+    paddingLeft: spacing(2),
+    paddingRight: spacing(2)
+  },
+  containerTablet: {
+    alignSelf: "center",
+    width: tabletWidth
   },
   datePublication: {
     ...fontFactory({
@@ -32,13 +31,16 @@ const sharedStyles = {
     color: colours.functional.secondary,
     marginTop: spacing(3)
   },
-  flag: {
-    flexDirection: "row",
+  flags: {
     marginBottom: spacing(3),
     marginTop: spacing(1)
   },
-  flagContainer: {
-    marginRight: spacing(3)
+  header: {
+    borderBottomColor: colours.functional.keyline,
+    borderBottomWidth: 1,
+    marginBottom: spacing(4),
+    paddingBottom: spacing(5),
+    paddingTop: spacing(7)
   },
   label: {
     marginBottom: spacing(2)