+++ packages/article-main-standard/src/styles/article-header/shared.js
@@ -2,13 +2,6 @@ import styleguide from "@times-components/styleguide";
 
 const { colours, fontFactory, spacing } = styleguide();
 const sharedStyles = {
-  articleFlag: {
-    flexDirection: "row",
-    marginBottom: spacing(2)
-  },
-  articleFlagContainer: {
-    marginRight: spacing(3)
-  },
   articleHeadLineText: {
     ...fontFactory({
       font: "headline",
@@ -26,6 +19,9 @@ const sharedStyles = {
   articleLabel: {
     paddingBottom: spacing(1),
     paddingTop: spacing(1)
+  },
+  flags: {
+    marginBottom: spacing(2)
   },
   headerTablet: {
     marginTop: spacing(6)