+++ packages/edition-slices/__tests__/ios/slices-with-style.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-slices-with-style.native";
+
+shared();