+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@0.7.0...@times-components/article-in-depth@0.7.1) (2019-01-24)
+
+
+### Bug Fixes
+
+* Versions due to multiple master merges ([#1617](https://github.com/newsuk/times-components/issues/1617)) ([9f732e2](https://github.com/newsuk/times-components/commit/9f732e2))
+
+
+
+
+
 # [0.7.0](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@0.6.0...@times-components/article-in-depth@0.7.0) (2019-01-23)