+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@0.6.0...@times-components/article-skeleton@0.6.1) (2019-01-24)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 # [0.6.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@0.5.0...@times-components/article-skeleton@0.6.0) (2019-01-23)