+++ packages/jest-configurator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.22](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@2.1.21...@times-components/jest-configurator@2.1.22) (2019-01-24)
+
+
+### Bug Fixes
+
+* Versions due to multiple master merges ([#1617](https://github.com/newsuk/times-components/issues/1617)) ([9f732e2](https://github.com/newsuk/times-components/commit/9f732e2))
+
+
+
+
+
 ## [2.1.21](https://github.com/newsuk/times-components/compare/@times-components/jest-configurator@2.1.20...@times-components/jest-configurator@2.1.21) (2019-01-22)
 
 **Note:** Version bump only for package @times-components/jest-configurator