+++ packages/slice-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.1.1...@times-components/slice-layout@0.1.2) (2019-01-24)
+
+**Note:** Version bump only for package @times-components/slice-layout
+
+
+
+
+
 ## [0.1.1](https://github.com/newsuk/times-components/compare/@times-components/slice-layout@0.1.0...@times-components/slice-layout@0.1.1) (2019-01-23)
 
 **Note:** Version bump only for package @times-components/slice-layout