+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.1](https://github.com/newsuk/times-components/compare/@times-components/utils@4.1.0...@times-components/utils@4.1.1) (2019-01-24)
+
+**Note:** Version bump only for package @times-components/utils
+
+
+
+
+
 # [4.1.0](https://github.com/newsuk/times-components/compare/@times-components/utils@4.0.19...@times-components/utils@4.1.0) (2019-01-23)